requirements.txt
crewai
langchain-google-genai
python-dotenv
akshare
pandas
duckduckgo-search
